Consolidate duplicate path handling functions
